move trackEvent out of LibraryCommander into LibraryAnalytics
