+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/LibraryAnalytics.scala
@@ -1,6 +1,7 @@
 package com.keepit.commanders
 
 import com.keepit.common.db.Id
+import com.keepit.common.mail.EmailAddress
 import com.keepit.model._
 import com.keepit.heimdal._
 import com.keepit.common.time._
@@ -11,6 +12,30 @@ import play.api.libs.concurrent.Execution.Implicits.defaultContext
 
 @Singleton
 class LibraryAnalytics @Inject() (heimdal: HeimdalServiceClient) {
+
+  def sendLibraryInvite(userId: Id[User], libId: Id[Library], inviteeList: Seq[(Either[Id[User], EmailAddress])], eventContext: HeimdalContext) = {
+    val builder = new HeimdalContextBuilder
+    builder.addExistingContext(eventContext)
+    builder += ("action", "sent")
+    builder += ("category", "libraryInvitation")
+    builder += ("libraryId", libId.id.toString)
+    builder += ("libraryOwnerId", userId.id.toString)
+    val numUsers = inviteeList.count(_.isLeft)
+    val numEmails = inviteeList.size - numUsers
+    builder += ("numUserInvited", numUsers)
+    builder += ("numNonUserInvited", numEmails)
+
+    heimdal.trackEvent(UserEvent(userId, builder.build, UserEventTypes.INVITED))
+  }
+
+  def acceptLibraryInvite(userId: Id[User], libId: Id[Library], inviteeList: Seq[(Either[Id[User], EmailAddress])], eventContext: HeimdalContext) = {
+    val builder = new HeimdalContextBuilder
+    builder.addExistingContext(eventContext)
+    builder += ("action", "accepted")
+    builder += ("category", "libraryInvitation")
+    builder += ("libraryId", libId.id.toString)
+    heimdal.trackEvent(UserEvent(userId, builder.build, UserEventTypes.INVITED))
+  }
 
   def followLibrary(userId: Id[User], library: Library, context: HeimdalContext): Unit = {
     val followedAt = currentDateTime
