+++ kifi-backend/modules/shoebox/app/com/keepit/commanders/LibraryCommander.scala
@@ -708,7 +708,7 @@ class LibraryCommander @Inject() (
       val (inv1, res) = successInvites.unzip
       inviteBulkUsers(inv1)
 
-      trackLibraryInvitation(inviterId, libraryId, inviteList.map { _._1 }, eventContext, action = "sent")
+      libraryAnalytics.sendLibraryInvite(inviterId, libraryId, inviteList.map { _._1 }, eventContext)
 
       Right(res)
     }
@@ -751,7 +751,7 @@ class LibraryCommander @Inject() (
         }
         val updatedLib = libraryRepo.save(lib.copy(memberCount = libraryMembershipRepo.countWithLibraryId(libraryId)))
         listInvites.map(inv => libraryInviteRepo.save(inv.copy(state = LibraryInviteStates.ACCEPTED)))
-        trackLibraryInvitation(userId, libraryId, Seq(), eventContext, action = "accepted")
+        libraryAnalytics.sendLibraryInvite(userId, libraryId, Seq(), eventContext)
         libraryAnalytics.followLibrary(userId, lib, eventContext)
         searchClient.updateLibraryIndex()
         Right(updatedLib)
@@ -1017,24 +1017,6 @@ class LibraryCommander @Inject() (
         }
       case _ => None
     }
-  }
-
-  private def trackLibraryInvitation(userId: Id[User], libId: Id[Library], inviteeList: Seq[(Either[Id[User], EmailAddress])], eventContext: HeimdalContext, action: String) = {
-    val builder = contextBuilderFactory()
-    builder.addExistingContext(eventContext)
-    builder += ("action", action)
-    builder += ("category", "libraryInvitation")
-    builder += ("libraryId", libId.id)
-
-    if (action == "sent") {
-      builder += ("libraryOwnerId", userId.id)
-      val numUsers = inviteeList.count(_.isLeft)
-      val numEmails = inviteeList.size - numUsers
-      builder += ("numUserInvited", numUsers)
-      builder += ("numNonUserInvited", numEmails)
-    }
-
-    heimdal.trackEvent(UserEvent(userId, builder.build, UserEventTypes.INVITED))
   }
 
   def convertPendingInvites(emailAddress: EmailAddress, userId: Id[User]) = {
